@@ -2,6 +2,8 @@
 
 namespace TheIconic\Tracking\GoogleAnalytics;
 
+use PHPUnit\Framework\TestCase;
+
 use TheIconic\Tracking\GoogleAnalytics\Parameters\ContentGrouping\ContentGroup;
 use TheIconic\Tracking\GoogleAnalytics\Parameters\EnhancedEcommerce\Affiliation;
 use TheIconic\Tracking\GoogleAnalytics\Parameters\EnhancedEcommerce\CouponCode;
@@ -28,7 +30,7 @@ use TheIconic\Tracking\GoogleAnalytics\Parameters\Hit\HitType;
  * Class AnalyticsTest
  * @package TheIconic\Tracking\GoogleAnalytics
  */
-class AnalyticsTest extends \PHPUnit_Framework_TestCase
+class AnalyticsTest extends TestCase
 {
     /**
      * @var Analytics
@@ -230,7 +232,7 @@ class AnalyticsTest extends \PHPUnit_Framework_TestCase
 
     public function testSendSimpleHit()
     {
-        $httpClient = $this->getMock('TheIconic\Tracking\GoogleAnalytics\Network\HttpClient', ['post']);
+        $httpClient = $this->createMock('TheIconic\Tracking\GoogleAnalytics\Network\HttpClient', ['post']);
 
         $this->analytics
             ->setProtocolVersion('1')
@@ -252,7 +254,7 @@ class AnalyticsTest extends \PHPUnit_Framework_TestCase
 
     public function testSendSimpleSslHit()
     {
-        $httpClient = $this->getMock('TheIconic\Tracking\GoogleAnalytics\Network\HttpClient', ['post']);
+        $httpClient = $this->createMock('TheIconic\Tracking\GoogleAnalytics\Network\HttpClient', ['post']);
 
         $this->analyticsSsl
             ->setAsyncRequest(true)
@@ -275,7 +277,7 @@ class AnalyticsTest extends \PHPUnit_Framework_TestCase
 
     public function testSendSimpleDebugHit()
     {
-        $httpClient = $this->getMock('TheIconic\Tracking\GoogleAnalytics\Network\HttpClient', ['post']);
+        $httpClient = $this->createMock('TheIconic\Tracking\GoogleAnalytics\Network\HttpClient', ['post']);
 
         $this->analytics
             ->setDebug(true)
@@ -298,7 +300,7 @@ class AnalyticsTest extends \PHPUnit_Framework_TestCase
 
     public function testSendBatchHits()
     {
-        $httpClient = $this->getMock('TheIconic\Tracking\GoogleAnalytics\Network\HttpClient', ['batch']);
+        $httpClient = $this->createMock('TheIconic\Tracking\GoogleAnalytics\Network\HttpClient', ['batch']);
 
         $this->analytics
             ->setDebug(true)
@@ -327,7 +329,7 @@ class AnalyticsTest extends \PHPUnit_Framework_TestCase
 
     public function testSendBatchHitsAfterEmpty()
     {
-        $httpClient = $this->getMock('TheIconic\Tracking\GoogleAnalytics\Network\HttpClient', ['batch']);
+        $httpClient = $this->createMock('TheIconic\Tracking\GoogleAnalytics\Network\HttpClient', ['batch']);
 
         $httpClient->expects($this->exactly(2))
             ->method('batch')
@@ -412,7 +414,7 @@ class AnalyticsTest extends \PHPUnit_Framework_TestCase
 
     public function testEmptyBatchHits()
     {
-        $httpClient = $this->getMock('TheIconic\Tracking\GoogleAnalytics\Network\HttpClient', ['batch']);
+        $httpClient = $this->createMock('TheIconic\Tracking\GoogleAnalytics\Network\HttpClient', ['batch']);
 
         $this->analytics
             ->setDebug(true)
@@ -463,7 +465,7 @@ class AnalyticsTest extends \PHPUnit_Framework_TestCase
 
     public function testFixTypos()
     {
-        $httpClient = $this->getMock('TheIconic\Tracking\GoogleAnalytics\Network\HttpClient', ['post']);
+        $httpClient = $this->createMock('TheIconic\Tracking\GoogleAnalytics\Network\HttpClient', ['post']);
 
         $this->analyticsSsl
             ->setUserTiminCategory('hehe')
@@ -500,7 +502,7 @@ class AnalyticsTest extends \PHPUnit_Framework_TestCase
             ->setClientId('666')
             ->setHitType('pageview');
 
-        $httpClient = $this->getMock('TheIconic\Tracking\GoogleAnalytics\Network\HttpClient', ['post']);
+        $httpClient = $this->createMock('TheIconic\Tracking\GoogleAnalytics\Network\HttpClient', ['post']);
 
         $httpClient->expects($this->once())
             ->method('post')
@@ -531,7 +533,7 @@ class AnalyticsTest extends \PHPUnit_Framework_TestCase
             ->setContentGroup('group', 1)
             ->setHitType('pageview');
 
-        $httpClient = $this->getMock('TheIconic\Tracking\GoogleAnalytics\Network\HttpClient', ['post']);
+        $httpClient = $this->createMock('TheIconic\Tracking\GoogleAnalytics\Network\HttpClient', ['post']);
 
         $httpClient->expects($this->once())
             ->method('post')
@@ -644,7 +646,7 @@ class AnalyticsTest extends \PHPUnit_Framework_TestCase
 
         $this->assertEquals('http://www.google-analytics.com/collect?v=1&tid=555&cid=666&aip=1&ds=call%20center&qt=560&ni=1&dclid=d_click_id&uip=202.126.106.175&ti=7778922&ta=THE%20ICONIC&tr=250&tt=25&ts=15&tcc=MY_COUPON&t=event&pa=purchase&cm3=50&pr1id=AAAA-6666&pr1nm=Test%20Product%202&pr1br=Test%20Brand%202&pr1ca=Test%20Category%203%2FTest%20Category%204&pr1va=yellow&pr1pr=50&pr1qt=1&pr1cc=TEST%202&pr1ps=2&pr2id=AAAA-5555&pr2nm=Test%20Product&pr2br=Test%20Brand&pr2ca=Test%20Category%201%2FTest%20Category%202&pr2va=blue&pr2pr=85&pr2qt=2&pr2cc=TEST&pr2ps=4&pr2cd1=iamcustomdim1&pr2cd2=iamcustomdim2&pr2cm1=666.99&pr2cm2=999&z=289372387623', $url);
 
-        $httpClient = $this->getMock('TheIconic\Tracking\GoogleAnalytics\Network\HttpClient', ['post']);
+        $httpClient = $this->createMock('TheIconic\Tracking\GoogleAnalytics\Network\HttpClient', ['post']);
 
         $httpClient->expects($this->once())
             ->method('post')
@@ -680,7 +682,7 @@ class AnalyticsTest extends \PHPUnit_Framework_TestCase
 
     public function testMinimumParametersForSendHitMissingClientIdButUserId()
     {
-        $httpClient = $this->getMock('TheIconic\Tracking\GoogleAnalytics\Network\HttpClient', ['post']);
+        $httpClient = $this->createMock('TheIconic\Tracking\GoogleAnalytics\Network\HttpClient', ['post']);
 
         $this->analytics
             ->setProtocolVersion('1')
@@ -701,7 +703,7 @@ class AnalyticsTest extends \PHPUnit_Framework_TestCase
 
     public function testMinimumParametersForSendHitWithClientIdButMissingUserId()
     {
-        $httpClient = $this->getMock('TheIconic\Tracking\GoogleAnalytics\Network\HttpClient', ['post']);
+        $httpClient = $this->createMock('TheIconic\Tracking\GoogleAnalytics\Network\HttpClient', ['post']);
 
         $this->analytics
             ->setProtocolVersion('1')
@@ -746,7 +748,7 @@ class AnalyticsTest extends \PHPUnit_Framework_TestCase
      */
     public function testSendPassesOptionsToHttpClient(array $options, array $expectedOptions, $async)
     {
-        $httpClient = $this->getMock('TheIconic\Tracking\GoogleAnalytics\Network\HttpClient', ['post']);
+        $httpClient = $this->createMock('TheIconic\Tracking\GoogleAnalytics\Network\HttpClient', ['post']);
 
         $analytics = new Analytics(false, false, $options);
         $analytics
